3nd commit - ready deploy

+++ src/Controller/FicheController.php
@@ -13,129 +13,186 @@ use App\Repository\FichePersoRepository;
 use App\Repository\RessourceRepository;
 use App\Repository\TypeInfoRepository;
 use App\Repository\UserRepository;
+use App\Service\SecurityService;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
+/**
+ * @Security(" (user.getActive() == true && user.getValidated() == true) || is_granted('ROLE_ADMIN') ")
+ */
 class FicheController extends AbstractController
 {
+    private $securityService;
+
+    public function __construct(SecurityService $securityService)
+    {
+        $this->securityService = $securityService;
+    }
+
+
     /**
      * @Route("/profil/{id}/fiche/list", name="fiche_list")
      */
     public function listFiche(int $id, UserRepository $userRepository){
 
         $user = $userRepository->find($id);
-        $fiches = $user->getFichePersos();
+        //récupération de l'instance du user pour contrôle sécurité
+        if ($this->getUser() == $user || $this->isGranted("ROLE_ADMIN")){
+            $fiches = $user->getFichePersos();
 
-        return $this->render('fiche/list.html.twig', [
-            'user'=>$user,
-            'fiches'=>$fiches
-        ]);
+            return $this->render('fiche/list.html.twig', [
+                'user'=>$user,
+                'fiches'=>$fiches
+            ]);
+        }
+        else{
+            //identification mauvais user + logMenace + alert + checkMenace + redirection
+            $realUser = $this->securityService->findRealUser();
+
+            return $this->redirectToRoute('fiche_list',[
+            'id' => $realUser->getId(),
+            ]);
+        }
     }
 
 
     /**
      * @Route("/fiche/{id}/detail", name="fiche_detail")
      */
-    public function ficheDetail(int $id, FichePersoRepository $fichePersoRepository, TypeInfoRepository $infoRepository, ChampsRepository $champRepository, RessourceRepository $ressourceRepository, Request $request): Response
+    public function ficheDetail(int $id, FichePersoRepository $fichePersoRepository, TypeInfoRepository $infoRepository, ChampsRepository $champRepository, RessourceRepository $ressourceRepository, UserRepository $userRepository, Request $request): Response
     {
         //récupération de l'instance de la fichePerso
         $fiche = $fichePersoRepository->find($id);
+        if(!$fiche){
+            $this->addFlash('warning', "Cette fiche n'existe pas" );
+            return $this->redirectToRoute('main_home');
+        }
 
-        //recupération liste types info
-        $listeInfo = $infoRepository->findAll();
+        //récupération de l'instance du user pour contrôle sécurité
+        $user = $fiche->getUser();
+        if ($this->getUser() == $user || $this->isGranted("ROLE_ADMIN")){
 
-        //récupération des champs et ressources associés à la fiche de perso triés par ordre (sort)
-        $champs = $champRepository->findBy(['fichePerso' => $id],['sort' => 'ASC']);
-        $ressources = $ressourceRepository->findBy(['fichePerso' => $id],['sort' => 'ASC']);
+            //recupération liste types info
+            $listeInfo = $infoRepository->findAll();
 
-        //récupération valeur bouton
-        $boutonFiche = $request->get("boutonFiche");
+            //récupération des champs et ressources associés à la fiche de perso triés par ordre (sort)
+            $champs = $champRepository->findBy(['fichePerso' => $id],['sort' => 'ASC']);
+            $ressources = $ressourceRepository->findBy(['fichePerso' => $id],['sort' => 'ASC']);
 
-        return $this->render('fiche/ficheDetail.html.twig', [
-            'fiche' => $fiche,
-            'listeInfo' => $listeInfo,
-            'champs' => $champs,
-            'ressources' => $ressources,
-            'boutonFiche' => $boutonFiche
-        ]);
+            //récupération valeur bouton
+            $boutonFiche = $request->get("boutonFiche");
+
+            return $this->render('fiche/ficheDetail.html.twig', [
+                'fiche' => $fiche,
+                'listeInfo' => $listeInfo,
+                'champs' => $champs,
+                'ressources' => $ressources,
+                'boutonFiche' => $boutonFiche
+            ]);
+        }
+        else{
+            //identification mauvais user + logMenace + alert + checkMenace + redirection
+            $realUser = $this->securityService->findRealUser();
+
+            return $this->redirectToRoute('fiche_list',[
+                'id' => $realUser->getId(),
+            ]);
+        }
     }
 
     /**
      * @Route("/profil/{id}/fiche/create", name="fiche_create")
+     *
      */
     public function ficheCreate(int $id, UserRepository $userRepository, ChampsRepository $champsRepository, RessourceRepository $ressourceRepository, Request $request): Response
     {
         //récupération de l'instance du user
         $user = $userRepository->find($id);
 
-        $fiche = new FichePerso();
-        $form = $this->createForm(FicheFormType::class, $fiche);
-        $form->remove('user');
-        $fiche->setUser($user);
+        //check si admin ou bon user
+        if ($this->getUser() == $user || $this->isGranted("ROLE_ADMIN"))
+        {
+            $fiche = new FichePerso();
+            $form = $this->createForm(FicheFormType::class, $fiche);
+            $form->remove('user');
+            $fiche->setUser($user);
 
-        $form->remove('nbChamps1');
-        $form->remove('nbChamps2');
-        $form->remove('nbChamps3');
-        $form->remove('nbChamps4');
-        $form->remove('nbChamps5');
-        $form->remove('nbChamps6');
-        $form->remove('nbChamps7');
+            $form->remove('nbChamps1');
+            $form->remove('nbChamps2');
+            $form->remove('nbChamps3');
+            $form->remove('nbChamps4');
+            $form->remove('nbChamps5');
+            $form->remove('nbChamps6');
+            $form->remove('nbChamps7');
 
-        $form->handleRequest($request);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager = $this->getDoctrine()->getManager();
-            try{
-                //gestion image
-                $image = $form->get('image')->getData();
-                if ($image){
-                    //genere un nouveau nom de fichier
-                    $fichier = md5(uniqid()).'.'.$image->guessExtension();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $entityManager = $this->getDoctrine()->getManager();
+                try{
+                    //gestion image
+                    $image = $form->get('image')->getData();
+                    if ($image){
+                        //genere un nouveau nom de fichier
+                        $fichier = md5(uniqid()) . '_Fiche.' . $image->guessExtension();
 
-                    //copie de la photo dans le dossier uploads
-                    $image->move($this->getParameter('images_directory'), $fichier);
+                        //copie de la photo dans le dossier uploads
+                        $image->move($this->getParameter('images_directory'), $fichier);
 
-                    //envoie du nom de fichier dans la BDD
-                    $fiche->setImage($fichier);
+                        //envoie du nom de fichier dans la BDD
+                        $fiche->setImage($fichier);
+                    }
+                    else{
+                        $fiche->setImage("icon_d20_mini.png");
+                    }
+
+                    //initialisation du nb de champs à 0 pour chaque type d'info
+                    $fiche->setNbChamps1(0);
+                    $fiche->setNbChamps2(0);
+                    $fiche->setNbChamps3(0);
+                    $fiche->setNbChamps4(0);
+                    $fiche->setNbChamps5(0);
+                    $fiche->setNbChamps6(0);
+                    $fiche->setNbChamps7(0);
+
+                    //initialisation du nb de ressource à 0
+                    $fiche->setNbRessource(0);
+
+                    $entityManager->persist($fiche);
+                    $entityManager->flush();
+                    $this->addFlash('success', 'Création de la fiche réussie !');
+
+                    $champs = $champsRepository->findBy(['fichePerso' => $id],['sort' => 'ASC']);
+                    $ressources = $ressourceRepository->findBy(['fichePerso' => $id],['sort' => 'ASC']);
+
+
+                    return $this->render('fiche/ficheDetail.html.twig',[
+                        'user' => $user,
+                        'fiche' => $fiche,
+                        'champs' => $champs,
+                        'ressources' => $ressources,
+                    ]);
                 }
-
-                //initialisation du nb de champs à 0 pour chaque type d'info
-                $fiche->setNbChamps1(0);
-                $fiche->setNbChamps2(0);
-                $fiche->setNbChamps3(0);
-                $fiche->setNbChamps4(0);
-                $fiche->setNbChamps5(0);
-                $fiche->setNbChamps6(0);
-                $fiche->setNbChamps7(0);
-
-                //initialisation du nb de ressource à 0
-                $fiche->setNbRessource(0);
-
-                $entityManager->persist($fiche);
-                $entityManager->flush();
-                $this->addFlash('success', 'Création de la fiche réussie !');
-
-                $champs = $champsRepository->findBy(['fichePerso' => $id],['sort' => 'ASC']);
-                $ressources = $ressourceRepository->findBy(['fichePerso' => $id],['sort' => 'ASC']);
-
-
-                return $this->render('fiche/ficheDetail.html.twig',[
-                    'user' => $user,
-                    'fiche' => $fiche,
-                    'champs' => $champs,
-                    'ressources' => $ressources,
-                ]);
+                catch(\Exception $e){
+                    $this->addFlash('warning', 'Erreur création de la fiche ! '.$e->getMessage());
+                }
             }
-            catch(\Exception $e){
-                $this->addFlash('warning', 'Erreur création de la fiche ! '.$e->getMessage());
-            }
+            return $this->render('fiche/create.html.twig', [
+                'user' => $user,
+                'ficheForm' => $form->createView()
+            ]);
         }
-        return $this->render('fiche/create.html.twig', [
-            'user' => $user,
-            'ficheForm' => $form->createView()
-        ]);
+        else{
+            //identification mauvais user + logMenace + alert + checkMenace + redirection
+            $realUser = $this->securityService->findRealUser();
+
+            return $this->redirectToRoute('fiche_create',[
+                'id' => $realUser->getId(),
+            ]);
+        }
     }
 
     /**
@@ -165,13 +222,17 @@ class FicheController extends AbstractController
                 if ($image){
                     //suppression de l'ancienne image
                     $oldImage = $fiche->getImage();
-                    if ($oldImage) {
-                        //suppression du fichier de l'ancienne photo
-                        unlink($this->getParameter('images_directory') . '/' . $oldImage);
+                    if ($oldImage && $oldImage != "icon_d20_mini.png") {
+                        try {
+                            //suppression du fichier de l'ancienne photo
+                            unlink($this->getParameter('images_directory') . '/' . $oldImage);
+                        } catch(\Exception $e){
+                            $this->addFlash('warning', 'Erreur suppression de l\'image ! ' . $e->getMessage());
+                        }
                     }
 
                     //genere un nouveau nom de fichier
-                    $fichier = md5(uniqid()).'.'.$image->guessExtension();
+                    $fichier = md5(uniqid()) . '_Fiche.' . $image->guessExtension();
 
                     //copie de la photo dans le dossier uploads
                     $image->move($this->getParameter('images_directory'), $fichier);
@@ -306,7 +367,7 @@ class FicheController extends AbstractController
         }
 
 
-        //formulaire ressource        (en cours)
+        //formulaire ressource
         if ($formRessource->isSubmitted() && $formRessource->isValid()) {
             try{
                 //gestion de l'ordre des ressources
@@ -399,7 +460,7 @@ class FicheController extends AbstractController
 
             //suppression du fichier image
             $image = $fiche->getImage();
-            if ($image) {
+            if ($image && $image != "icon_d20_mini.png") {
                 //suppression du fichier de l'ancienne photo
                 unlink($this->getParameter('images_directory') . '/' . $image);
             }
